# open_api/swagger.yaml
basePath: /
definitions:
  dto.FileResponse:
    properties:
      content_type:
        type: string
      created_at:
        type: string
      id:
        type: string
      link:
        type: string
      name:
        type: string
      size:
        type: integer
      updated_at:
        type: string
    type: object
  echo.HTTPError:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
  description: 'File storage service: uploads go to S3, metadata to Postgres, downloads
    via presigned links.'
  title: Bucket Inventory API
  version: "1.0"
paths:
  /files:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.FileResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: List files with presigned download links
      tags:
      - files
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Upload a file
      tags:
      - files
  /files/{id}:
    delete:
      parameters:
      - description: File ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Delete a file
      tags:
      - files
    get:
      parameters:
      - description: File ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get a file with a presigned download link
      tags:
      - files
swagger: "2.0"